fix: cast to numeric for ROUND() in PostgreSQL migration 096

+++ backend/migrations/096_upstream_markup_percentage.sql
@@ -3,10 +3,10 @@
 -- 旧值 1.3（1.3x 倍率）→ 新值 30（加价 30%）
 
 -- upstream_sites 表
-UPDATE upstream_sites SET price_multiplier = ROUND((price_multiplier - 1) * 100, 2) WHERE price_multiplier != 1;
+UPDATE upstream_sites SET price_multiplier = ROUND(((price_multiplier - 1) * 100)::numeric, 2) WHERE price_multiplier != 1;
 UPDATE upstream_sites SET price_multiplier = 0 WHERE price_multiplier = 1;
 ALTER TABLE upstream_sites ALTER COLUMN price_multiplier SET DEFAULT 0;
 
 -- upstream_managed_resources 表（0 仍然表示"继承站点默认"）
-UPDATE upstream_managed_resources SET price_multiplier = ROUND((price_multiplier - 1) * 100, 2) WHERE price_multiplier > 0 AND price_multiplier != 1;
+UPDATE upstream_managed_resources SET price_multiplier = ROUND(((price_multiplier - 1) * 100)::numeric, 2) WHERE price_multiplier > 0 AND price_multiplier != 1;
 UPDATE upstream_managed_resources SET price_multiplier = 0 WHERE price_multiplier = 1;
